requirements.txt
torch>=2.0.0
numpy>=1.21.0
pyyaml>=6.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0
wandb>=0.15.0
tqdm>=4.64.0
sentencepiece>=0.1.97
einops>=0.6.0
pytest>=7.0.0